+++ RayTracerChallengev3/src/DataStructs/Point4.h
@@ -4,9 +4,9 @@ namespace RayTracer
 {		
 	struct Point4 : public Vector4
 	{
-		Point4() : Vector4(0.0f, 0.0f, 0.0f, 1.0f) {}
-		Point4(float lx, float ly, float lz) : Vector4(lx, ly, lz, 1.0f) {}
-		Point4(const Vector4& v) : Vector4(v.x, v.y, v.z, 1.0f) {}
+		Point4() : Vector4(0.0, 0.0, 0.0, 1.0) {}
+		Point4(double lx, double ly, double lz) : Vector4(lx, ly, lz, 1.0) {}
+		Point4(const Vector4& v) : Vector4(v.x, v.y, v.z, 1.0) {}
 	};
 }
