+++ RayTracerChallengev3/src/Geometry/Material.h
@@ -6,13 +6,13 @@ namespace RayTracer
 	{	
 		Color color;
 		IPattern* pattern;
-		float ambient;		// 0-1 - light reflected from other objects in environment (phong treats as constant)
-		float diffuse;		// 0-1 - light reflected from matte surface
-		float specular;		// 0-1 - light source reflection
-		float shininess;	// ~10-~200 - reflection of the light source itself (small # large highlight, large # small highlight)
-		float reflective;	// 0 non reflective, 1 mirror
-		float transparency; // 0-1 - 0=opaque, 1=transparent
-		float refractiveIndex; // the degree that light will bend when entering or exiting a material
+		double ambient;		// 0-1 - light reflected from other objects in environment (phong treats as constant)
+		double diffuse;		// 0-1 - light reflected from matte surface
+		double specular;		// 0-1 - light source reflection
+		double shininess;	// ~10-~200 - reflection of the light source itself (small # large highlight, large # small highlight)
+		double reflective;	// 0 non reflective, 1 mirror
+		double transparency; // 0-1 - 0=opaque, 1=transparent
+		double refractiveIndex; // the degree that light will bend when entering or exiting a material
 
 		// example refractiveIndex values
 		// vacuum 1
