+++ RayTracerChallengev3/src/Geometry/Ray.h
@@ -10,7 +10,7 @@ namespace RayTracer
 		Ray(const Point4& lOrigin, const Vector4& lDirection) : origin(lOrigin), direction(lDirection) {}
 		Ray(const Ray& copy) : origin(copy.origin), direction(copy.direction) {}
 
-		Point4 position(const float t) const;
+		Point4 position(const double t) const;
 		
 		Ray transform(const Matrix4x4& matrix) const;
 	};
