+++ RayTracerChallengev3/src/Patterns/GradientPattern.cpp
@@ -7,7 +7,7 @@ namespace RayTracer
 	Color GradientPattern::patternAt(const Point4& p) const
 	{
 		Color distance = color2 - color1;
-		float fraction = p.x - std::floorf(p.x);
+		double fraction = p.x - std::floor(p.x);
 
 		return color1 + (distance * fraction);
 	}
