+++ RayTracerChallengev3/src/Shapes/Cube.h
@@ -6,13 +6,13 @@ namespace RayTracer
 	{
 		Point4 origin;
 
-		Cube() : origin(Point4(0.0f, 0.0f, 0.0f)) { }
+		Cube() : origin(Point4(0.0, 0.0, 0.0)) { }
 		Cube(const Point4& lOrigin) : origin(lOrigin) { }
 
 		Vector4 localNormalAt(const Point4& localPoint) const;
 
 		void localIntersectBy(const Ray& localRay, std::vector<Intersection>& intersections) const;
 
-		void checkAxis(const float& axisOrigin, const float& rayDirection, float& tmin, float& tmax) const;
+		void checkAxis(const double& axisOrigin, const double& rayDirection, double& tmin, double& tmax) const;
 	};
 }
